kripto-hans-job-service/requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
iniconfig==1.1.1
kombu==5.1.0
logassert==6
packaging==21.0
pluggy==0.13.1
prompt-toolkit==3.0.20
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
